requirements.txt
# Dependências Python para processamento de PDFs e geração de testes

# Processamento de PDFs
PyMuPDF>=1.23.0          # fitz - Extração de texto de PDFs
pdf2image>=1.16.0        # Conversão PDF -> imagens
pytesseract>=0.3.10      # OCR - Reconhecimento de texto em imagens
Pillow>=10.0.0           # Manipulação de imagens

# Utilitários
python-dateutil>=2.8.0   # Manipulação de datas

# Opcional: Para desenvolvimento
# pytest>=7.0.0           # Testes unitários
# black>=23.0.0           # Formatação de código
# pylint>=2.17.0          # Análise estática
